+++ docs/source/changelog.rst
@@ -12,6 +12,7 @@ Version 0.3
  * ``WEBLAB_CALLBACK_URL`` is now optional, and ``/callback`` by default.
  * Added ``current_task`` that can be called inside a task to get the ``task_id`` or update data.
  * Added ``current_task.data`` and ``current_task.update_data`` so as to update JSON-friendly data to measure the progress of the task.
+ * Added ``@weblab.user_loader``. If set, you can later run say ``user = weblab_user.user``, and it returns a user (e.g., from your database)
 
 Version 0.2
 -----------